cf: read api_token & account_id from a config file
Switches from using an API Key (which has full perms) to an
API Token (which can be limited to just certain permissions).
Reads such a token from a file (which will go in puppet-private soon;
there's already such a token in pwstore).

Depends upon a spicerack with I98adc429f7.

Change-Id: I4654649d3efe52c99bfd489059d2f39872aa1ca8