sre.wdqs.data-transfer: fix syntax, simplify rule
We were missing both commas and semicolons. We can also simplify the two rules to one.

NOTE: This was tested on wdqs1009 and appears to be generating the
expected iptables rule. For example with a source of
`'wdqs1008.eqiad.wmnet'` we end up with the following iptables rule:

ACCEPT     tcp  --  wdqs1008.eqiad.wmnet  anywhere             tcp dpt:9876

Note: It appears that if ferm cannot resolve a hostname then it silently
fails rather than producing an error.

Bug: T206951
Change-Id: Ia7dc920e7797def13f3e5f24e1773bea7ca63f3d