elasticsearch: Let spicerack handle wait for all write queues to clear
WHY THIS CHANGE WAS MADE

Running a rolling upgrade failed because the previous query was
returning [], the empty list, on eqiad. codfw gave the right response
but we had eqiad hardcoded.

CHANGES

(1) Delegate the "wait for all write queues to be empty" logic to spicerack, which
itself includes a change to amend the prometheus query to reflect the new state of things

(2) expose a --write-queue-datacenters flag to specify a list of one or
more dcs to check rather than all of them; this flag should be used to override when
there's a problem with the metrics in a given datacenter, not used routinely

Bug: T261239
Depends-On: I0147a638b1d001e4ca25b97002968fe87dfc8107
Change-Id: I2fcc6ae26dbb1e940ace609b12f40f9d8cde466b