sre.hosts.reboot-single: refine the icinga check/repool logic
We really don't want to automatically pool a server that has services
that are not ok on icinga. On the other hand, sometimes icinga is
super slow in rescheduling checks after a reboot, so we need to force
checks when a server has been rebooted and after the first puppet run
not all checks have recovered.

Change-Id: I659de42369d0649547e93d8ed5a8cf2bdd674cef