Introduce sre.host.reboot-cluster
This cookbook is intended to be used for rolling reboots of clusters
that are part of an LVS load-balanced pool.

It will split the whole cluster in subgroups, depool all the servers
in that group, reboot them, and repool them.
It will also wait for the first puppet run to happen, and
for icinga to go all green before repooling. It can optionally ignore
the icinga failures, as we've noticed sometimes icinga is too slow in
rescheduling checks.

Some of the logic here could proably be generalized in spicerack's
LBRemoteCluster class, by generalizing around the run() method. I
deemed more important to have a working implementation now, so that we
know better what to generalize in spicerack later.

Change-Id: Ie1e2a84510d6e447299939b432de2905c7fb842c