sre.switchdc.services: use configuration file
We needed to refresh the list of services we will switch over, and it
was a good time to also start reading the values from a configuration
file, that is generated by puppet.

This way, we won't have to update this recipe further for any service
that is added to discovery.

Change-Id: I1e63b063fd2d24f6c5e385a2bfb4443424c798fd