switchdc: Run Puppet on DB masters after setting read-write
In the 2020-09-01 switchover we saw some spurious alerting noise, as DB
masters didn't immediately pick up the change in which DC was expected
to be read-write. We may or may not address this differently in the long
run, but at least for the 2020-10-27 switchback, let's add this puppet
run as an easy workaround.

Bug: T261767
Change-Id: If36abc5d801aabefae4be38b57c346fb5c2b3b75