sre.switchdc.mediawiki: Add a step to restart Envoy on jobrunners.
In the last switchover, we found that jobrunners in DC_FROM were
continuing to try to edit, reporting errors from the DB master
(correctly) rejecting edits due to being in read-only mode.

The cause was that changeprop maintains a long-lived connection to the
jobrunners. When we updated the jobrunners' DNS record to point to
DC_TO, changeprop didn't notice, and continued using its existing
connection to jobrunners in DC_FROM.

With this change, we restart Envoy (running as a TLS terminator on the
server side) on all jobrunners in DC_FROM after the switchover. When
Envoy shuts down, changeprop closes its long-running connection and
opens a new one. Because it re-resolves the jobrunner DNS entry in the
process, it picks up the change and starts talking to DC_TO.

Change-Id: Ibf5160b6b75b7ebfce46b0165e775ce26b9add09