00-warmup-caches: Repeat until execution time converges.
Instead of running a flat three times, we tend to re-run this section
until it stops getting any faster, so let's automate that.

Separately we'll be doing some further work on the warmup script,
including adding a broader set of URLs, warming up the api servers as
well as appservers, and maybe rewriting the whole thing in Python
instead of NodeJS -- so it's possible we'll move the rerun logic into
the script and take it back out of the cookbook -- but in the meantime
this is a good incremental improvement.

Change-Id: If9f7acc914d21f945157c83b001aa742be58cb5e