sre.hadoop.change-distro-from-cdh: use confirm_on_failure()
The confirm_on_failure function wraps a function call offering a
ask_confirmation prompt if an exception is raised. This is useful for
this cookbook since we will deal with a lot of hosts, and we don't want
to stop the procedure if any minor hiccup occurs (networking blip, single
host failures, etc..).

Change-Id: I188dd6984d45b138f93b29afaba121b267e11a3b