Add cookbook to upgrade hadoop client nodes to Bigtop
This cookbook is a companion to the change-distro-from-cdh one.
Once the Hadoop cluster is upgraded, the Analytics team will need
to upgrade all the satellite "client" nodes as well, that are a
lot (~30 nodes). This cookbook should allow the team to upgrade
node be node without any manual intervention, to avoid mistakes.
It also allows to selectively upgrade groups of nodes to ease the
rollout.

Bug: T269919
Change-Id: If96020642f9ff4198585ea9fc8b3ab9d4f893994