postgres: use remote script on replica to resync
Currently we chain commands here to attempt replica creation. These
commands don't currently work for a few reasons (the data directory is
not recreated after it is removed, the recovery file being absent
means the replica will not be able to start replication from the
master and will require a resync), so as we fix them it makes more
sense to move this logic to a script on each replica.

Bug: T275381
Change-Id: I78fc04c6a44ef6aa2bedc803c59f31d648c77582