wmcs.toolforge.etcd: Added cookbook to depool and remove a node
This is a minimal cookbook that allows to porperly depool and remove
an etcd node from a toolforge installation.

Requires some spicerack patches to be merged and released, mainly:
https://gerrit.wikimedia.org/r/c/operations/software/spicerack/+/666961/2

And some cumin patches not released yet, so to test you need to
download latest master and install locally.

This is meant to be run from your laptop, the config used is:
cookbook.yaml
```
    ---
    cookbooks_base_dir: ~/Work/wikimedia/operations-cookbooks
    instance_params:
        cumin_config: cumin-config.yaml
```

cumin-config.yaml
```
    ---
    transport: clustershell
    log_file: cumin.log
    default_backend: direct
    clustershell:
        ssh_options:
            # needed for vms that repeat a name
            - |
                -o StrictHostKeyChecking=no
                -o "UserKnownHostsFile=/dev/null"
                -o "LogLevel=ERROR"
```

Bug: T274497
Change-Id: I6ece0975cc569ff728f2975f1373ccb23460f3f9
Signed-off-by: David Caro <david@dcaro.es>