sre.ganeti.makevm: Automatically generate fqdn from hostname
To prevent accidents where the fqdn doesn't match the datacenter, we
should autogenerate the fqdn based on the user's networking and datacenter
preferences.

This cookbook now takes a hostname (e.g. "test1001"), and then builds the
fqdn and prompts the user for confirmation. The numerical part will also
be checked to ensure it starts with the right number, if it exists.

Since users may still try to pass in a fqdn, an error will be raised if
the provided hostname as a "." in it.

Bug: T276516
Change-Id: I98831f840bcffcf7fd3c0c1533139038d95028c0