tests: fix pip backtracking
With the current setup of minimizing the number of different
virtualenvs used by tox we ended up hitting an issue of pip
backtracking. As prospector seems to be the most likely culprit here
because has a lot of dependencies, and in the past too we had issues
between prospector and flake8 dependencies, move prospector to its own
virtualenv.

This is entirely copied from I0bcbb448ef by Volans.

Change-Id: I86d2b790ac4c08a8ca5a5044b2a34f8043ab53d9