wmcs.toolforge.etcd: make sure the etcdctl node is not the new one
As we are sorting alphabetically, 10 goes before 1 so if the new node
is number 10, it will be picked as the node to do etcdctl commands on,
but it might not be ready yet.
This checks that the selected node is not the new one, and if so picks
the next in the pool.

Bug: T267082
Change-Id: I6f580b828a3140246c39e39fa034c6991455cc5e
Signed-off-by: David Caro <david@dcaro.es>