elasticsearch: refactor various rolling operations
There was a lot of duplication between the various cookbooks. The core
pattern is the same between each operation, with small differences for a
service restart vs full reboot and whether we are doing a plugin
upgrade.

Condense the rolling [restart, reboot] x [upgrade, don't upgrade] into
one generic "rolling-operation" cookbook. This new cookbook will always
at least perform a restart, but can optionally perform a plugin upgrade
and/or a full reboot.

Bug: T280221
Change-Id: I04106008e0d10a831beeea0b7f32e354bbbec8d3