wmcs.ceph: Added mon upgrade cookbooks
This adds a simple set of cookbooks to roll-upgrade the mons on a ceph
cluster.

It includes:

* Setting the cluster in maintenance mode (noout, norebalance).
* Unsetting the cluster maintenance mode (noout, norebalance).
* Upgrade a singe ceph host (works the same for all).
* Upgrade all the mon hosts in a cluster.

It has into account that the upgrade to octopus will trigger a new set
of health warnings, so it will consider the upgrade successful even if
those warnings appear.

Tested on codfw, where the current mon nodes (cloudcephmon2*) were
upgraded with it successfully.

Signed-off-by: David Caro <david@dcaro.es>
Change-Id: Icb76d788672cb37f0d518c6861d6d67582d37d36