00-reduce-ttl: Sleep after updating TTL
Now that the warmup is faster than five minutes, let's add this to the
cookbook so the operator doesn't have to keep track of the time. It's a
little too bad that means we can't run 00-warmup-caches in parallel, but
not a big deal.

In case of an emergency, the operator can Ctrl+C after seeing the
sleep message since it's the last thing.

Change-Id: I9dc5ebb63ec8407325afe293eca44cbd72e7bcb9