sre.switchdc.mediawiki: Run the warmup cache script at least 6 times
During the June 2021 DC switchover we ended up running the script twice
because we didn't feel that the cluster was sufficiently warm after the
first run. In theory the convergence logic should take care of all of
that, but it doesn't appear to do so so far.

Adding a minimum number of iterations is straightforward to do and costs
us very little. This is a 00 step, so it's not time sensitive. The main
win is that it's significantly less likely for the cookbook operator to
have to analyze the warmup results and consider doing another run.

6 was chosen because the script is usually converging after 3 loops,
so twice that.

Bug: T285802
Change-Id: Ib08dccaf951281a39d3ed7d5100093225402c237