sre.ores.roll-restart-workers: move to LBConfig
The new LBConfig logic is not 100% equal to the previous one,
but it should be fine for the ORES use case. For example,
there is no real need to sleep between uwsgi and celery restarts
when the host is depooled.
Also move the list of ORES clusters to __init__.py.

Change-Id: Ief7819b91bb43d76225750059a7bef49e9848b52