Elasticsearch cookbooks: Represent ops as enum
Introduce an enum for the various possible operations. This seems
slightly cleaner than having a boolean flag for each potential
operation.

Validate that there is either a reboot or restart operation, because we
stop the ES process(es) by default, so without a reboot or restart we
would not start Elasticsearch up again.

Bug: T280221
Change-Id: Idaa28b4199f50fffa7dccd30bb3851dec851e734