sre.switchdc.mediawiki: Downtime read-only checks on the DB primaries
Also remove the downtimes post-switchover, after running Puppet to update
the expected state.

The DB primaries all have Icinga alerts to verify that they're
read-write in the active DC and read-only in the passive, but we can't
update the expected state at the same instant as we flip RO state, so
the alerts would page during the switchover if not downtimed.

DBAs currently do this manually. Now that Spicerack supports downtiming
services in Icinga, we can automate it.

Bug: T285803
Change-Id: I40c25cecdce67b645180f36a322afe8cd1635ddf
Depends-On: I2025adb157e4f494c3b82a196416c038417d2087