start_instance_with_prefix: work around extra stderr message
Turns out that on debian stretch, the default root profile has a line
at the end of the profile (`msg n || true`) that outputs a line on
stderr, and cumin/spicerack are unable to distinguish stderr from
stdout, making any method dependent on the output (ex. reboot) fail
due to being unable to parse it.

This tries to alleviate that issue by changing the profile file right
after the instance is spun up.

Change-Id: I7defa1e95a7ec7363d24217803bc8ce624c20c4c
Signed-off-by: David Caro <dcaro@wikimedia.org>