upgrade-varnish: support frontend instance only
Since varnish is now only deployed as a frontend instance there is no
backend service to take care of (varnish.service). There is also no need
to test if the backend is serving requests properly, since it is not
varnish anymore but rather ATS.

Additionally, move the cookbook under sre.cdn as this is not a generic
cookbook to be run on any host.

Change-Id: I376d84f8b8f261eb1b8d0f9ab1297e567bf97b19