elastic: relax & restore perms during upgrade
Refactor to use a single operation, rather than
a list.

--

Previously we've used this cookbook with the --upgrade
flag in order to upgrade the wmf-elasticsearch-search-plugins,
but now we're also going to use it to upgrade the elasticsearch
package itself. As such we need to give the elasticsearch user
the ability to write to the relevant /etc/elasticsearch directories.

From
https://www.elastic.co/guide/en/elasticsearch/reference/7.17/elasticsearch-keystore.html#keystore-upgrade

> When Elasticsearch is installed from a package manager,
an upgrade of the on-disk keystore to the new format
is done during package upgrade. In other cases,
Elasticsearch performs the upgrade during node startup.
This requires that Elasticsearch has write permissions
to the directory that contains the keystore.

Bug: T301955
Change-Id: Ib421f989461ee04d2ab19b10efb857ffece0a4b7