elastic: Bring back stopping new replicas during restart
I suggested removing this along with removing the bits to stop writes
in 60123a3296. After further review of elastic docs and monitoring a
restart of cloudelastic it seems like we should have only removed the
frozen indexes part, which requires extensive application support, and
kept the bits to prevent new replicas from being created while waiting
for nodes to restart.

Change-Id: If7fc70f8c112d4bc1a3611cf6959ba845a4586fa