elastic: allow waiting for yellow instead of green
When performing an elasticsearch upgrade, if a primary shard is
scheduled onto an upgraded node, the replica shards must be scheduled to
nodes of equal or higher version. In order to proceed with the upgrade
operation, we need to allow a special condition where instead of waiting
for green status, we allow yellow status if there's no relocating or
initializing shards.

See
https://www.elastic.co/guide/en/elastic-stack/current/upgrading-elasticsearch.html

Note: this relies on functionality added in
https://gerrit.wikimedia.org/r/c/operations/software/spicerack/+/776999/

Bug: T304570
Depends-On: I5a9eaf2c86865bb5b6dfff49e7bb3de46bee3241
Change-Id: I8bf2273097bd124f821aa2b4e1b41ce23f1f4a98