Modify Ganeti addnode.py script function when detecting bridge status
The is_valid_bridge function evaluates if MAC addresses are being
learnt correctly on the external intefaces of every bridge defined.
This validates that the switch the host is connected to is properly
configured to trunk all required Vlans to the Ganeti host.

The previous config did this sucessfully when the device was a vlan
sub-interface, but it failed for the 'private' bridge as the member
port is the physical interface itself, and thus has no '@' symbol in
the iproute2 command output.

This fairly ugly change should ensure it works in both cases from here
out.

Change-Id: I979d6c568e6a625b4e2d9b0f64e4716c0457cfb8