Update bridge detection code
Replace the brctl check (which is dependent on the interface name)
with a check based on "ip", which also has the benefit of no longer
depending on bridge-utils (which will go away when we move to systemd-
networkd).

Also move the check into the more logical place, the recently added
is_valid_bridge() function.

Change-Id: I808c431bf722ec62ded0ae2e98d3446d00e27196