Validate the Ganeti node has been added to Hiera (and thus Ferm)
When a new Ganeti node is added to a cluster, it also needs to be added to
the Hiera settings (which is used to generate the Ferm rules). Validate
that the new node is present in the Ferm config on the master node, if
it's missing that means that various master commands towards the host
would fail to connect.

Change-Id: I866027e0a9768f795ff4d24056afe661b3b606f6