elastic: log return value of reimage cookbook
We're not going to try to handle all the possible corner cases; rather
we're just going to let the cookbook keep running as normally, and then
the operator can come clean up later.

Basically the cookbook's either going to keep running until it finishes
the cluster, or it's going to fail on a subsequent batch.

Bug: T308606
Change-Id: I3799ae3e56a6b8b20736b791979714ca7b85051f