wmcs: Parse enums at argparse level
This makes it "easy" to parse enums at the argparse level, instead of
deferring the parsing later in the code.

DebianVersion had to be handled differently as it uses the lowercase
of the name instead of the value of the enum entries.

Also fixed a couple issues created by the type confusion.

Change-Id: I92df7ec165ab21df8e2c65b9d75071b57d6f7270
Signed-off-by: David Caro <dcaro@wikimedia.org>