elastic: don't disable replica alloc for reimage
Stopping replica shard allocation is a good idea during restarts or
similarly "short" duration operations to force Elasticsearch to wait and
recover the shards from disk rather than immediately starting a network
recovery while the host(s) are restarting.

With a re-image, we're wiping away the data so there's no local recovery
to perform anyway. Furthermore, reimages take 45-60 minutes which is a
little longer than we'd like to wait even if the data were still there.

--

As far as implementation, we use ExitStack [1] which allows us
to enter another context manager. This lets us actually check if we're
in a reimage operation before we enter the
self.elasticsearch_clusters.stopped_replication context manager.

[1]
https://docs.python.org/3/library/contextlib.html#contextlib.ExitStack

Bug: T289135
Change-Id: I9ae82256a995c665813be1bbc5cd3c3f64acad16