global: add inventory module
This changes the concept of 'deployment' in favor of a wider one,
usable for other clusters than openstack and changes the --deployment
to --cluster-name.

Functionality wise, the only change is for the ceph set/unset
maintenance cookbooks, where instead of the controller node, now we
can pass the cluster name.

Bug: T314870
Change-Id: I6a21b218d9133c60b4d76aa2121ad215a598534d
Signed-off-by: David Caro <dcaro@wikimedia.org>