Add cookbook to restart pybal
This cookbook simply does as follows:
* restart the service on one host at a time
* check that all the bgp sessions are established again

In the future we can make it smarter by checking preemptively
which hosts can be restarted at the same time, if we really
feel the need, but it seemed like a premature optimization right now.

The reason why we're not using thanos here is that we need a more
fine-grained time resolution and the fact we can't fully trust the timestamps
in thanos responses.

Change-Id: I7e30d6d17957d740e8ca9d8745286c2d5a7c5a00