wdqs: auto-extract kafka timestamp
The streaming updater needs to be fed a kafka
timestamp to know "when" to start reading
the log of change events from.

Previously this was being fed in manually by
operators via CLI input.

That timestamp is available in the first lines
of the TTL dumps, so let's grab it automatically.

Also validate timestamp is in the right iso format.

Bug: T325114
Change-Id: Icef40bdb5dfa96c3f6918bab2a16e40833e5d2dd