wdqs: make depool the default behavior
Before this patch, we had to pass --depool to
tell it to depool the underlying host;
however our usage comment didn't include the depool flag.

Given that we want to depool in all cases except the
rare case where we're reloading on a test host,
let's make it the default and require an explicit
--no-depool instead.

Usage comment updated accordingly.

A SMALL NOTE ON HISTORICAL CONTEXT
We've generally done reloads on the test host and then
done a data-transfer to the actual production hosts, so
the old default does make some sense in that respect.
However we don't have a test host for codfw, so overall
it's probably wisest to default to the depool.

Also apply naming convention changes for
data-transfer and restart cookbooks.

Bug: T323096
Change-Id: I6bb9e36bb60204136f622442f41b8766c1e97051