Allow the wikireplicas.add-wiki cookbook to replace existing views
In some cases the views for databases in the wikireplacas already
exist, whilst the other steps in the cookbook are required. In these
cases the cookbook fails because maintain-views prompts for input and
cannot receive it.

Adding the --replace-all argument to the maintain-views command prevents
the maintain-views script for prompting for permission to overwrite the
existing views.

Bug: T310872
Change-Id: Ia9671d666970bf443ab39482a6b955501b44c401