Add sre.k8s.upgrade-cluster
This cookbook should automate the reimage of etcd, control-plane
and workers of a k8s cluster ease the upgrade to a new version.
The procedure is aimed to automate what listed in T326340.
It it still not perfect but it removes a lot of burdern from
the operator's time, especially for big clusters.

Since __init__.py was added, prospector started to complain
about the reboot nodes cookbook. Fix errors as well to allow
CI to pass.

Bug: T327767
Change-Id: I892962e5ae1163265bdf0747b73b496b44c04502