sre.gitlab.upgrade: check for unknown version
GitLab API version endpoint returns "'unknown' instead of raising an error.
This change checks for a unkown version to catch invalid api calls (bad token or url).

This change also moves pre-loading of gitlab-ce apt packge to the beginning
of the cookbook to make sure it runs before a full backup.

Bug: T323569
Change-Id: I4290b18ed66f533961613b05554d4db29cd2b289