sre.discovery.datacenter: add --fast-insecure switch for pool/depool
Using the cookbook right now takes about 20 minutes, as the time spent
migrating services scales linearly with the number of services.

While proceeding in this reasonably secure manner makes a lot of sense
for programmatic depooling, it's not what we want if we are acting in an
emergency.

To allow the cookbook to run faster, add a --fast-insecure switch that
will:
* First change the pooled state of all service in the datacenter as
  usual
* Then, instead of checking and wiping cache for each individual service,
  wipe the cache of the whole discovery.wmnet subdomain twice, 30
  seconds apart

This should significantly speed up the process, because we're skipping
waiting for a service state change to propagate before moving to the
next one. we rather take for granted that eventually gdnsd will have the
right entries for all services, so just wiping the cache should be enough.

Change-Id: I4dbb86d64131a37b9b34fa635dd7b33df25f35c0