sre.puppet.sync-netbox-hiera: Use netbox GraphQL endpoint to fetch data
Update the current cookbook to directly fetch the netbox data using the
GraphQL endpoint.  This allows us to keep all code in one place and not
split between the netbox-extra and this repo.

However so far i dont see any real speed improvement:
  $ time sudo cookbook -vvvv  -c /home/jbond/cookbook.yaml --dry-run \
      sre.puppet.sync-netbox-hiera 'testing GraphQL'
  real    0m59.435s
  user    0m2.491s
  sys     0m0.382s

  $ time sudo cookbook --dry-run sre.puppet.sync-netbox-hiera \
      'testing GraphQL'
   real    1m0.915s
   user    0m2.381s
   sys     0m0.367s

I have tested this with dry-run mode and the resulting diff is empty

Change-Id: Icd3212529ad8e56a7b29db3e948d920040f24814