sre.k8s.upgrade-cluster: fix alertmanager's catch all
The original intent was to downtime the whole k8s Prometheus
instance to avoid alarms firing during the run of the cookbook.
The dry-run didn't catch the problem, but AlertManager
returns an HTTP error for the ".*" selector, so the best
compromise seems to fall back to the affected node's nodeset.

Bug: T327767
Change-Id: Iccb404a9e83c9cbf620e55fc7bc002d16b66355e