sre.k8s.upgrade-cluster: fix puppet disable action
In the prepare_nodes function I used "disabled()" that returns
a context manager (to be used in "with" blocks for example)
rather than the "disable()" function, so no puppet disable
action was triggered.

Also re-enable puppet in case etcd wipe only is used
(otherwise a reimage will automatically re-enable it).
The etcd sanity checks should be performed also when
a wipe happens (not strictly needed but better safe
than sorry).

Bug: T327767
Change-Id: I8ae76ad7b5e1f1c21402473d20af571b93d83028