sre.k8s.upgrade-cluster: simplify etcd cluster procedure
I ran the cookbook one time and the etcd upgrade procedure
has been really painful. The current workflow is:
- stop puppet and etcd on all nodes.
- reimage one node at the time.
The idea was that the first node would have bootstrapped
a single-node cluster, and the others would have joined
as soon as reimages completed. We use discovery SRV
records for etcd and this seems to complicate things, since
etcd tries to do leader election etc.. even when there
is only a single node up.
To avoid these problems, I propose to move to a simpler
procedure:
- wipe the cluster from current data.
- reimage one etcd node at the time, keeping the cluster up.

Bug: T327767
Change-Id: I76a0116553e6700532d913280303789790714021