sre.k8s.upgrade-cluster: simplify code and extend downtimes
The exception handling code for remote code execution errors
can be simplified with wmflib's confirm_on_failure, as Riccardo
noted in a previous code review.

Extend also the downtimes for the affected nodes, since the cookbook
is likely to last a long time and they risk to expire sooner
than expected.

Bug: T327767
Change-Id: Idf67ad08913d8382faa261b99c1d345f086d7f73