k8s.wipe-cluster: Don't disable puppet or downtime etcd
The cookbook does not make any changes to etcd nodes (it's just
interacting with etcd) so there is no need to disable puppet or downtime
etcd.

Also the + operator is not supported for NodeSets
(affected_nodes.update) and I've removed the second ask_confirmation
before the wipe as the use already confirmed in line 124.

Change-Id: I424d222cd3823406f2f549ee1fdda2246db34a2b