sre.hardware.upgrade-firmware: sync firmware store when downloading files
Update the upgrade cookbook to ensure we sync the firmware store when
ever we download a new file.  I have opted to sync the entire store
instead of the individual files to ensure we pick up anything manually
copied.  Further if we run on every file download the end result should
mostly be that we only sync one file.

Bug: T332158
Change-Id: I8869866afafafc7ca9ad3ca7d97ce17c933df34a