sre.hosts.reboot-single.py: replace "pool" with "depool"
While running the reboot-single cookbook (called from the upgrade
firmware one) I got this error: https://phabricator.wikimedia.org/P45904

The end result was that the reboot didn't happen, and the upgrade
firmware cookbook kept checking Redfish for updates.

IIUC the reference to "pool" is wrong, and causing another exception
that breaks the logic of the cookbook.

Bug: T325153
Change-Id: I51e1b5bb326f463684b12cc87c63f40e60fb2b1c