sre.swift.remove-ghost-objects: new cookbook
This is a cookbook to handle the problem of "ghost" objects - entities
that exist in some container dbs and thus at least sometimes appear in
container listings, but do not in fact exist.

We need to clear them out so that our container listings actually
match their contents, and so that our rclone sync script will run
successfully (it errors when trying to copy or delete ghost objects).

Listing a container with ghosts in is unreliable - some quantity of
the ghosts will be returned in each listing, stochastically. So
instead we collect the on-disk databases for the container in
question, and inspect those.

This cookbook has already been tested with some containers highlighted
by rclone job failures.

Bug: T327253
Change-Id: Ia717d3e6f253f7c1890643ad2e57ff173e0f2acd