sre.hardware.upgrade-firmware: update to only upgrade one hardware type
Update the cookbook so that we only try to update one hardware type per
run.  For network cards we look for a port with a link and choose that
card.  For storage controllers we choose the ones named raid.  in both
situations if there are more then one then the user will pick

Change-Id: I7fe5a3d1269f5fd5885faa413d0c7e72c802889b