Improve logic getting switch port when primary IP is on bridge device
This change addresses the original issue I opened the parent phab
task for, using the improved data model in Netbox to reliably get
the physical port that belongs to the primary device if it is a
bridge.

With our current host configs I don't think the current work-around
could fail.  But I can imagine some alternate scenarios we could
have where it may not be reliable.  This improves it by ensuring we
never pick a port that doesn't belong to the primary bridge.

Bug: T296832
Change-Id: Ia5efb1c02bdc76a93af21335c086fbafae3f5ccf