sre.ganeti.makvm: update the default memory to 1.5
Also update the logic so the minimum allowed is 1.5G.  A bug in
bookwork[1] prevents d-i from working with less then 1.5 GB of memory.
As such change the default and minimum allowed value.

We could have added logic to only impose the 1.5 limit on bookworm
machines whoever I think going forward this is likely to be the default
and I don't think it would do to much harm giving other machines and
additional 500m of memory  but please say if this assumption flawed

Change-Id: Icbdf2692a686bc8479469757047f15272401f745