sre.wdqs.data-transfer: fix broken logic
We recently refactored this to allow a wdqs-all option for
blazegraph_instance that tells the cookbook to transfer *both*
categories and wikidata (two separate blazegraph instances). We
neglected to consider the difference in behavior however when the new
value of args.blazegraph_instance is taken into account.

Also when we're passing the --force flag which deletes already extant
.jnl files, let's also wipe away the data_loaded file (which serves to
indicate to the wdqs updater that a successful data transfer has been
completed and therefore the updater is safe to start)

Change-Id: I3d9a2180bc181dfce78951ab5d85ff062de113d2
Bug: T321605