Add cookbook to manage users SSH keys on SONiC devices
With only 2 SONiC devices and SSH key support through their API on their
roadmap, this cookbook worksaround the current limitation.
The idea is to keep managing users through Homer, but having to run that
cookbook when ssh keys change. It could eventually be run as a cron job
if the keys change often (not the case actually), or we could implement
the "all" selector once we have more than 2 devices.

Not ideal but the least worse option for something hopefully temporary.

Bug: T338028
Change-Id: If23c01c34127fe3a14c6533a6c4e891e30646832