Allow disabling puppet on reboot
To disable Puppet during a reboot, one would need to use pre_action()
and post_action() with manual disabling. Unfortunately, the
post_action() would never execute as _reboot_action required a
successful Puppet run before execution, causing an infinite waiting
period.

This will allow a simple bool to specify whether or not to disable
Puppet in the same way as disable_puppet_on_restart. It defaults to
False as that's the incumbent behavior.

Bug: T342182
Change-Id: I750958940f3f349bfd711df9f3139bfa8a28f09a