sre.dns: add a new cookbook for durum reboot/service restarts
This commit introduces a new cookbook for the rolling service restarts
and host reboots of durum, the check service for Wikimedia DNS. This is
inspired by and copied from the Wikimedia DNS cookbook in the same path,
with changes for durum.

The nginx -> anycast-hc -> bird binding was already added to the durum
hosts and this cookbook takes that into account.

Change-Id: I7d734f4fe7daf6b775dc90ff50044cc0b6debc9b