mysql: Stop removing the downtime after clone is done
It takes some time for the replica to catch up from the new primary and
the current cookbook happily removes the downtime afterwards leading to
a page to SREs like what just happened.

Instead of using context manager that would remove it, just downtime it
and let the DBA decide what to do afterwards.

Change-Id: I491bc281045215e8b0f2d3f10c220cfe42b4a7cf