[sre.kafka] Use broker in-sync status as a gate between broker restarts
The way the cookbook used to work was by restarting a broker, waiting
a fixed amount of time, and moving on to the next one.
This has 2 main downsides:
- if the broker catches up in less than this sleep time, this introduced
  un-necessary delay in the rolling-restart process
- more dangerously, if the broker hasn't caught up after the delay,
  then another broker would be restarted, potentially destabilizing the
  cluster.

We rely on the `kafka-broker-in-sync` command introduced in
https://gerrit.wikimedia.org/r/c/operations/puppet/+/959162 to exit with
status 0 if all partitions assigned to the broker are in-sync, and with
status 1 if not.

I also took this as an opportunity to convert both restart/reboot kafka
cookbooks to the `SREBatchRunner` structure, which should give us
reproducibility between cookbooks.

The new API is now
```
% cookbooks sre.kafka.roll-restart-reboot-brokers \
	--alias kafka-jumbo-eqiad \
	--reason 'reload new config' \
	restart_daemons
```
to rolling-restart the cluster and
```
% cookbooks sre.kafka.roll-restart-reboot-brokers
	--alias kafka-jumbo-eqiad \
	--reason 'upgrade kernel' \
	reboot
```
to rolling-reboot

Bug: T346741
Change-Id: I6d9b47a93f384a987d3a0a5bd07bea804fe86ca5