[gitlab/upgrade] Add option to skip backups on replica upgrades
When we upgrade gitlab, we take a backup on each of the replica hosts.
This takes upwards of 90 minutes right now, meaning we're spending 3
hours backing up the backup hosts.

This adds an option to skip creating a backup, but won't allow it to be
set if the host is running the 'gitlab.wikimedia.org' instance.

Change-Id: Ief399ba868273e9343a4217b41c37bebf381b9b0