Adjust reimage cookbook config for DHCP binding clear workaround
While testing for codfw I realised that the reimage cookbook was
not issuing the command to clear DHCP bindings / mac-ip cache
for servers in codfw rows A-B connected to the new switches there.

Ultimately DHCP option 97 is the answer here.  For now we can check
if the server is connected to an effected switch based on the site,
switch model, and presence (or not) of a dedicated rack-vlan
matching our naming convention.

Bug: T306421
Change-Id: I48f12bcea3dd55af33636687cd9e48d64768771b