sre.puppet-migrate-*: update cookbooks to stop puppet
Due to the fact that puppet-run has some natural splay there is still a
bit of a race condition.  As such we need to code so that we first stop
puppet-agent-timer.service to ensure any jobs that are started are stopped

We also change how we disable the puppet-agent-timer.timer.  previously
we simply stopped this.  however the next puppet run would enable it
which is not desirable.  as such we touch a file which is used by the
systemd timer ConditionPathExists=!/run/puppet/disabled see dependent cr

Depends-On: I56f0fcbe9ebdf2568eef74d7ad2a7370b2054d58
Change-Id: I7c58c2ddf11db0dba73a92224ba6a63cc08abd09