sre.puppet.migrate-*: allow some steps to fail
The most common thing to fail in this cookbook is the installation of
puppet-agent.  this most often happens if a user forgets to merge the
yaml change or make some mistake/typo.  As such simply retrying the
install is not very useful.  instead let the function fail.
in the next step we check that we are at the correct version and if
not we throw an exception and give the user an option to retry.

This seems more desirable as it gives the user a chance to check there
yaml.

Change-Id: Iadb33558367df0ce199821f63c4f67a0d5aa3569