Add basic validation to Junos config command execution flow
This patch changes the run_junos_commands() function for network
cookbooks to check that the "commit complete" line is present
after we issue the "commit confirmed" command.  If not it will
raise an exception.

This improves on the previous behaviour in which the cookbook
blindly continues if the commit fails.

Bug: T353825
Change-Id: I2d6db748e4764c270a8d74dbb5c4bf6816da3ce8