sre.mysql.clone: Silence SQL injection warning
There are no inputs that affect this query, and certainly no non-trusted
inputs. This is a false positive.

Adding a nosec marker to L177 is impossible as that's a multiline
string, and Bandit does not have have a function to disable something on
the next line. Thankfully moving the .replace() to a separate statement
is enough to confuse Bandit and make the warning go away.

Change-Id: I09eae7c7b869267b0d58c4cf1925a9d3b90eee08