Adjust reimage cookbook to clear switch caches for vms too
We now have Ganeti hosts connected to switches running EVPN, which
are susceptible to a bug caused by the switch DHCP snooping, and
the fact it clears the ARP cache (but not mac-ip table) when a
DHCP lease expires.

The way we currently deal with this is to issue the command to delete
the DHCP binding after the reimage.  This triggers the bug, but then
we immediately clear the mac-ip table, which allows normal ARP to
happen again.

This patch adjust the function that does this to also take action if
the device being reimaged is a VM.  There are some quirks due to how
our ganeti hosts move the primary IP to a non-physical device but
logic is the same.

Tested and working with a test vm.  Will re-test a physical before
merge.

Bug: T306421
Change-Id: I39477c13cc6875e0a06de2df7cf3c105be339686