[gitlab] Pause/Prompt before restarting gitlab during upgrade
When performing the upgrade, we often take a full backup. This can take
over an hour, and last a variable amount of time. Since we have no way
of deterministically saying when it will happen, it's hard to give our
users adequate notice. This way we can pause, notify users, and then
allow the restart to contiune.

Change-Id: I9b5bab8c0b5a1a37a5238150234381cba5aab83b